Start working on high level api
